requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.3.4
    # via django
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
dj-database-url==0.5.0
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-model-utils==4.1.1
    # via -r requirements.in
django==3.2.2
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-model-utils
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.6.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
gunicorn==20.1.0
    # via -r requirements.in
idna==2.10
    # via requests
psycopg2-binary==2.8.6
    # via -r requirements.in
pyjwt==2.1.0
    # via djangorestframework-simplejwt
python-decouple==3.4
    # via -r requirements.in
pytz==2021.1
    # via django
requests==2.25.1
    # via -r requirements.in
sqlparse==0.4.1
    # via django
urllib3==1.26.4
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
